Moved relay modern into src folder
It complained that Watchman wasn't installed (even though wasn't relay-compiling with --watch). Needed to update relay packages and move run relay-compile against sub folder (https://github.com/facebook/relay/issues/2042)
