+++ NavigationReact/sample/relay-modern/package.json
@@ -17,13 +17,13 @@
         "navigation-react": "*",
         "react": "*",
         "react-dom": "*",
-        "react-relay": "^1.3.0",
-        "relay-compiler": "^1.3.0",
+        "react-relay": "^1.4.1",
+        "relay-compiler": "^1.4.1",
         "webpack": "^3.5.5"
     },
     "scripts": {
-        "start": "babel-node ./server.js",
+        "start": "babel-node ./src/server.js",
         "updateSchema": "babel-node ./updateSchema.js",
-        "relay": "relay-compiler --src . --schema ./schema.json"
+        "relay": "relay-compiler --src src --schema ./schema.json"
     }
 }
